requirements.txt
ipython
pandas
seaborn
scikit-learn
sentence-transformers
llama-index
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-vector-stores-duckdb
FlagEmbedding
llama-index-postprocessor-flag-embedding-reranker
llama-index-embeddings-ollama
squarify
pypalettes